+++ laktory/models/resources/databricks/__init__.py
@@ -24,6 +24,7 @@ from .mwsnetworkconnectivityconfig import MwsNetworkConnectivityConfig
 from .mwspermissionassignment import MwsPermissionAssignment
 from .notebook import Notebook
 from .notificationdestination import NotificationDestination
+from .obotoken import OboToken
 from .permissions import Permissions
 from .pipeline import Pipeline
 from .pythonpackage import PythonPackage