+++ laktory/models/stacks/stack.py
@@ -37,6 +37,7 @@ from laktory.models.resources.databricks.notebook import Notebook
 from laktory.models.resources.databricks.notificationdestination import (
     NotificationDestination,
 )
+from laktory.models.resources.databricks.obotoken import OboToken
 from laktory.models.resources.databricks.permissions import Permissions
 from laktory.models.resources.databricks.pipeline import Pipeline as DatabricksPipeline
 from laktory.models.resources.databricks.pythonpackage import PythonPackage
@@ -151,6 +152,7 @@ class StackResources(BaseModel):
     databricks_networkconnectivityconfig: dict[str, MwsNetworkConnectivityConfig] = {}
     databricks_notebooks: dict[str, Notebook] = {}
     databricks_notificationdestinations: dict[str, NotificationDestination] = {}
+    databricks_obotokens: dict[str, OboToken] = {}
     databricks_permissions: dict[str, Permissions] = {}
     databricks_pythonpackages: dict[str, PythonPackage] = {}
     databricks_queries: dict[str, Query] = {}