+++ tests/resources/test_accesscontrolruleset.py
@@ -1,0 +1,10 @@
+from laktory.models.resources.databricks import AccessControlRuleSet
+
+acrs = AccessControlRuleSet(
+    name="test",
+    grant_rules=[{"role": "some_role", "principals": ["p0", "p1"]}],
+)
+
+
+def test_accesscontrolruleset():
+    assert acrs.name == "test"